requirements.txt
pymysql~=1.1.0
python-dotenv~=1.0.0
Flask~=3.0.0
cryptography~=41.0.5